requirements.txt
flask==2.3.3
flask-cors==4.0.1
python-dotenv==1.0.0
PyJWT==2.9.0
requests==2.32.3
httpx>=0.24.0
gunicorn==21.2.0
azure-core==1.35.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-search-documents==11.5.3
azure-ai-documentintelligence==1.0.2
azure-cosmos==4.9.0
openai==1.93.0
pandas==2.3.1
scikit-learn>=1.0
PyMuPDF>=1.18.0
tiktoken
langchain
tabulate
